possible we need to use context for things like biome since its weird to drill it so much. but idk if its worth it yet. If i get more of these top-down drillers i'll probably build out a context

+++ public/generator/jsonProcessor.py
@@ -12,7 +12,8 @@ def text_to_json():
     # Clear out the processed folder or make it if it doesn't exist
     if os.path.exists(processed_directory):
         for file in os.listdir(processed_directory):
-            os.remove(os.path.join(processed_directory, file))
+            if file.endswith('.json'):
+                os.remove(os.path.join(processed_directory, file))
     else:
         os.makedirs(processed_directory)
